requirements.txt
opencv-python
pyzbar
Pillow
numpy
qrcode